feat and fix: OptionContainer. Score and buttons.
I was changing the background color of all wrong options when selecting one of them, so I added the selectedOption state to change the color only of the selected question. The correct question is now displayed in green, and the wrong one in red. Additionally, I added the score to enhance the user experience, along with a button to restart the quiz in case they score less than 50 points and a button to return to the home page if they score more than 50 points.